Prevent Linux crash when keyring is locked
When the Linux login keyring is locked when the app is opened
and the user clicks "Cancel" on the resulting prompt,
`keytar.FindCredentials` (within the `keytar` node native module)
segfaults, as the value of `secret` is NULL. This leads to a total
Electron crash and resulting core dump.

Short-circuit the process if secret is returned NULL.

This issue has been
[reported upstream](atom/node-keytar#477)
with a corresponding fix, but GitHub has [opted to sunset Atom and
related projects](https://github.blog/2022-06-08-sunsetting-atom/).